Automatic formation forming kind of works...
The main problem seems to be that if an imp in the middle of the
formation leaves, then the indexes get messed up...

In particular, if the indexes are:
0 1 2 3

And imp 2 leaves, then the indexes are:
0 1 3

Then if a new imp joins, the indexes are:
0 1 3 3

And so there are duplicate indexes, which is bad.

I suppose the indexes need to somehow be handed out differently.
Maybe just have a queue of indices?? When an imp leaves, they just
add their index back into the queue.